+++ src/commands/__init__.py
@@ -4,7 +4,7 @@ reload("src.commands", ["package_report", "override_report", "diff_report",
 	"refresh_report", "diff_single", "toggle_override", "create_override",
 	"diff_override", "edit_override", "delete_override", "freshen_override",
 	"diff_package", "diff_externally", "revert_override", "freshen_package",
-	"modify_mark"])
+	   "modify_mark", "context_create_override"])
 
 from .package_report import OverrideAuditPackageReportCommand
 from .override_report import OverrideAuditOverrideReportCommand
@@ -12,6 +12,7 @@ from .diff_report import OverrideAuditDiffReportCommand
 from .diff_single import OverrideAuditDiffSingleCommand
 from .toggle_override import OverrideAuditToggleOverrideCommand
 from .create_override import OverrideAuditCreateOverrideCommand
+from .context_create_override import OverrideAuditContextCreateOverrideCommand
 from .diff_override import OverrideAuditDiffOverrideCommand
 from .revert_override import OverrideAuditRevertOverrideCommand
 from .diff_externally import OverrideAuditDiffExternallyCommand
@@ -33,6 +34,7 @@ __all__ = [
 	# Override commands
 	"OverrideAuditToggleOverrideCommand",
 	"OverrideAuditCreateOverrideCommand",
+	"OverrideAuditContextCreateOverrideCommand",
 	"OverrideAuditDiffOverrideCommand",
 	"OverrideAuditDiffExternallyCommand",
 	"OverrideAuditRevertOverrideCommand",