+++ src/commands/create_override.py
@@ -4,7 +4,7 @@ import sublime_plugin
 import os
 
 from ..core import diff_override, packages_with_overrides, log
-from ..core import PackageListCollectionThread
+from ..core import setup_new_override_view, PackageListCollectionThread
 from ..browse import ResourceType, PackageResourceBrowser
 
 
@@ -18,14 +18,23 @@ class OverrideAuditCreateOverrideCommand(sublime_plugin.WindowCommand):
 	editing if it exists or create a buffer and set up so that saving will
 	create an override on save.
 	"""
-	def run(self, package=None, file=None):
+	def run(self, package=None, file=None, include_existing=False):
 		# If a packge or file name is missing, prompt for the given item and
 		# reinvoke ourselves. When given a file but no package, the file is
 		# ignored and the user will be prompted to pick the file.
 		if package is None or file is None:
+			if include_existing:
+				res_type = ResourceType.ALL
+				annotate = True
+				p_filter = lambda p: not p.is_disabled
+			else:
+				res_type = ResourceType.NONOVERRIDE
+				annotate = False
+				p_filter = lambda p: bool(p.package_file()) and not p.is_disabled
+
 			return PackageResourceBrowser(package, file, self.window,
-				ResourceType.NONOVERRIDE, unknown=False,
-				p_filter=lambda p: bool(p.package_file()),
+				res_type, unknown=False, annotate_overrides=annotate,
+				p_filter=p_filter,
 				on_done=lambda p,r: self.pick(p, r)).browse()
 
 		# Open normally if an unpacked copy exists; fallback for manual calls
@@ -40,10 +49,9 @@ class OverrideAuditCreateOverrideCommand(sublime_plugin.WindowCommand):
 
 		self.window.run_command("open_file", {"file": "${packages}/" + res})
 
-		# Mark the view as a potential new override and set it up.
+		# Get the active view and set it up as a potential new override.
 		view = self.window.active_view()
-		view.settings().set("_oa_is_new_override", True)
-		self.setup_view(view)
+		setup_new_override_view(view)
 
 	def pick(self, pkg_info, resource):
 		if pkg_info is not None:
@@ -51,23 +59,6 @@ class OverrideAuditCreateOverrideCommand(sublime_plugin.WindowCommand):
 				"package": pkg_info.name,
 				"file": resource
 			})
-
-	def setup_view(self, view):
-		if view.is_loading():
-			return sublime.set_timeout(lambda: self.setup_view(view), 10)
-
-		settings = sublime.load_settings("Preferences.sublime-settings")
-		mini_diff = settings.get("mini_diff")
-
-		# File is left as a scratch buffer until the first modification
-		view.run_command("move_to", {"to": "bof"})
-		view.set_read_only(False)
-
-		# Sublime turns off mini_diff for packed files that it opens.
-		if mini_diff:
-			view.settings().set("mini_diff", mini_diff)
-			reference_doc = view.substr(sublime.Region(0, len(view)))
-			view.set_reference_document(reference_doc)
 
 
 ###----------------------------------------------------------------------------